refactor(architecture): break the attention cycle trio (MOD-AI-SESSION-ATTENTION wave)
Three cuts, one direction rule: the attention pipeline may depend on the
session core, never the reverse.

1. The attention session-key codecs sink into the kernel
   (src/attentionSessionKeys.ts) so runtime settlement and workspace
   projections stop edging back into the attention module; attentionProject
   re-exports them for existing readers.
2. hasWorkspaceRuntimeContinuity (src/runtimeOwnership.ts) joins the kernel;
   it is shared workspace-runtime infrastructure consumed by four modules.
3. src/aiSessions/statusCapability.ts is re-owned by the attention module
   (it drives the attention dot and the evaluation queue), and the runtime
   lifecycle settlement primitives move into runtimeSettlementCapability.ts
   where their only consumer lives.

Measured: the three ATTENTION 2-cycles (provider, runtime, workspace)
dissolve; baseline 9 -> 6 fingerprints; ARCH-WAIVER-001 keeps only the two
non-attention pairs. SCC membership is unchanged cluster-level debt
(ARCH-WAIVER-004); per policy it blocks whole-program acceptance, not
strict mode. New characterization tests pin the session-key encodings.
No behavior change.
