+++ .ci/architecture-debt-baseline.json
@@ -5,9 +5,6 @@
   "rules": {
     "module-cycle": {
       "fingerprints": [
-        "2:MOD-AI-SESSION-ATTENTION->MOD-AI-SESSION-PROVIDER",
-        "2:MOD-AI-SESSION-ATTENTION->MOD-AI-SESSION-RUNTIME",
-        "2:MOD-AI-SESSION-ATTENTION->MOD-WORKSPACE-IDENTITY",
         "2:MOD-AI-SESSION-PROVIDER->MOD-AI-SESSION-RUNTIME",
         "2:MOD-AI-SESSION-PROVIDER->MOD-DASHBOARD-SHELL",
         "2:MOD-AI-SESSION-PROVIDER->MOD-WORKSPACE-IDENTITY",
