+++ src/aiSessions/runtimeSettlementCapability.ts
@@ -2,15 +2,8 @@
 
 import type * as vscode from 'vscode';
 import type { AiSessionProviderId } from '../models';
-import {
-    runAiSessionRuntimeLifecycleTask,
-    settleAiSessionRuntimeLifecycles,
-} from './attentionController';
-import type {
-    AiSessionAttentionEvaluation,
-    AiSessionRuntimeLifecycleCandidate,
-} from './attentionController';
-import { getAttentionRuntimeSessionKey } from './attentionProject';
+import type { AiSessionAttentionEvaluation } from './attentionController';
+import { getAttentionRuntimeSessionKey } from '../attentionSessionKeys';
 import { cloneAiSessionRuntimeIdentity } from './runtimeTypes';
 import type { AiSessionRuntimeSnapshot } from './runtimeTypes';
 import type AiSessionTerminalService from './terminalService';
@@ -229,4 +222,114 @@ export function createAiSessionRuntimeSettlementCapability(
             }
         },
     };
+}
+export interface AiSessionRuntimeLifecycleCandidate {
+    key: string;
+    sessionKey?: string;
+    state: 'completed' | 'stopped';
+}
+
+export type AiSessionRuntimeLifecycleFailureOperation = 'evaluate' | 'release';
+
+export interface SettleAiSessionRuntimeLifecyclesOptions<TCandidate extends AiSessionRuntimeLifecycleCandidate> {
+    candidates: readonly TCandidate[];
+    evaluateAttention: () => Promise<AiSessionAttentionEvaluation>;
+    release: (candidate: TCandidate) => void | Promise<void>;
+    reportFailure?: (
+        operation: AiSessionRuntimeLifecycleFailureOperation,
+        category: 'unexpected',
+        key: string | undefined
+    ) => void;
+}
+
+export interface AiSessionRuntimeLifecycleSettlementResult {
+    releasedKeys: string[];
+    retainedKeys: string[];
+}
+
+export function runAiSessionRuntimeLifecycleTask(
+    operation: string,
+    task: () => unknown | Promise<unknown>,
+    reportFailure: (operation: string, category: 'unexpected') => unknown | Promise<unknown>
+): Promise<void> {
+    return Promise.resolve().then(task).then(() => undefined, async () => {
+        try {
+            await reportFailure(operation, 'unexpected');
+        } catch (_reportError) {
+            // A diagnostic reporter must not escape the safe lifecycle boundary.
+        }
+    });
+}
+
+export async function settleAiSessionRuntimeLifecycles<
+    TCandidate extends AiSessionRuntimeLifecycleCandidate
+>(
+    options: SettleAiSessionRuntimeLifecyclesOptions<TCandidate>
+): Promise<AiSessionRuntimeLifecycleSettlementResult> {
+    const candidates = deduplicateLifecycleCandidates(options.candidates);
+    let evaluation: AiSessionAttentionEvaluation;
+    try {
+        evaluation = await options.evaluateAttention();
+    } catch (_error) {
+        options.reportFailure?.('evaluate', 'unexpected', undefined);
+        return {
+            releasedKeys: [],
+            retainedKeys: candidates.map(candidate => candidate.key).sort(),
+        };
+    }
+
+    const inScope = new Set(evaluation.inScopeSessionKeys);
+    const candidateSessionKey = (candidate: TCandidate): string => candidate.sessionKey || candidate.key;
+    const overflowed = new Set(evaluation.overflowedSessionKeys);
+    const hasAttentionEvidence = (candidate: TCandidate): boolean => {
+        const sessionKey = candidateSessionKey(candidate);
+        return (evaluation.eventIdsBySession[sessionKey] || []).length > 0
+            || overflowed.has(sessionKey);
+    };
+    const safeToRelease = candidates.filter(candidate => {
+        if (candidate.state === 'stopped' || !evaluation.enabled) {
+            return true;
+        }
+        const sessionKey = candidateSessionKey(candidate);
+        if (!inScope.has(sessionKey) || !hasAttentionEvidence(candidate)) {
+            return true;
+        }
+        return evaluation.published;
+    });
+
+    const eligibleByKey = new Map<string, TCandidate>();
+    for (const candidate of safeToRelease) {
+        eligibleByKey.set(candidate.key, candidate);
+    }
+    const releasedKeys: string[] = [];
+    for (const candidate of [...eligibleByKey.values()].sort((left, right) =>
+        left.key.localeCompare(right.key))) {
+        try {
+            await options.release(candidate);
+            releasedKeys.push(candidate.key);
+        } catch (_error) {
+            options.reportFailure?.('release', 'unexpected', candidate.key);
+        }
+    }
+    const released = new Set(releasedKeys);
+    return {
+        releasedKeys,
+        retainedKeys: candidates.map(candidate => candidate.key)
+            .filter(key => !released.has(key)).sort(),
+    };
+}
+
+function deduplicateLifecycleCandidates<TCandidate extends AiSessionRuntimeLifecycleCandidate>(
+    candidates: readonly TCandidate[]
+): TCandidate[] {
+    const byKey = new Map<string, TCandidate>();
+    for (const candidate of candidates) {
+        if (candidate && candidate.key && (candidate.state === 'completed' || candidate.state === 'stopped')) {
+            const existing = byKey.get(candidate.key);
+            if (!existing || (existing.state === 'stopped' && candidate.state === 'completed')) {
+                byKey.set(candidate.key, candidate);
+            }
+        }
+    }
+    return [...byKey.values()];
 }
